requirements.txt
pytorch-transformers==1.1.0
torch==1.3.1
tqdm==4.36.1
stanfordcorenlp==3.9.1.1
spacy==2.1.9
rouge==0.2
rouge-score==0.0.3
language-evaluation==0.1.0
seqeval==0.0.12
